requirements.txt
flask
flask-restx
prometheus_client
pyyaml
pandas
reportlab
gunicorn
waitress
python-dateutil
